+++ CHANGELOG.md
@@ -11,7 +11,10 @@ You can also check the
 
 # Unreleased
 
+- Features
+  - Limit values displayed in the legend now include the unit if it's available
 - Fixes
+  - Column line combo charts now work correctly with temporal entity X dimension
   - Default min custom scale domain value now defaults to 0 if no negative
     values are present
   - Chart area no longer overflows below X or Y axes when setting custom scale